frwdWord: search, context mark, set searchforward
v:searchforward being set is required to ensure correctness of typical
n/N usage.

context mark is set for convenience of end-user: After a bunch of hops,
the user should be able to just `` or '' back to their original
position.

search() is just used because I think using functions for that is
cleaner.