doc/stim.txt: Updated to new method and mappings.
It's possibly a bit verbose, and maybe should be in tagged, per-map
descriptions similar to vim's itself.

Also adjusted the suggested mapping in section 3, as I personally feel
that clearing matches is not really in the scope of StIm, and hence I
don't think it should be suggested to the user (especially if they don't
know what clearmatches() does).